requirements.txt
boto3==1.4.1
kafka-python==1.3.5
psycopg2==2.6.2
python-dateutil==2.6.1
retrying==1.3.3
msgpack==0.5.1
urllib3==1.22
